requirements.txt
Django==4.2.1
django_environ==0.10.0
google_api_python_client==2.86.0
google_auth_oauthlib==1.0.0
google_auth==2.19.0
google-auth-httplib2==0.1.0